added force_virtualgl option to headless anatomist
this option is set by default, it fixes cases when Xvfb reports an existing GLX extension, but which is not actually working.

The xvfb server shutdown has been revised: instead of using an atexit() function, which may not be called in case of crash, use POpen preexec_fn so setup a signal to be sent to the child (Xvfb) when the parent process (anatomist) dies.